refactor: modularize pushup detector
